demo/requirements.txt
gradio>=4.0.0
torch>=2.4.0
transformers>=4.45.0
peft>=0.13.0
accelerate>=1.0.0
bitsandbytes>=0.44.0
pillow>=10.0.0
imageio[pyav]>=2.30.0
numpy>=1.24.0
